Introduce Wozmon
The source file is compiled using CA65 and linked into a bin using
LD65, which are both part of the CC65 compiler toolchain. The
compilation code is very simple and can be easily run in a terminal
prompt, however I wrote a Makefile for it as well (traditionally,
GNU Make is *NIX, however there exists MinGW and NMake also works).